+++ src/onegov.pdf/onegov/pdf/__init__.py
@@ -1,6 +1,8 @@
 from onegov.pdf.page_functions import page_fn_footer
 from onegov.pdf.page_functions import page_fn_header
 from onegov.pdf.page_functions import page_fn_header_and_footer
+from onegov.pdf.page_functions import page_fn_header_logo
+from onegov.pdf.page_functions import page_fn_header_logo_and_footer
 from onegov.pdf.pdf import Pdf
 from onegov.pdf.signature import LexworkSigner
 
@@ -9,6 +11,8 @@ __all__ = [
     'LexworkSigner',
     'page_fn_footer',
     'page_fn_header_and_footer',
+    'page_fn_header_logo_and_footer',
+    'page_fn_header_logo',
     'page_fn_header',
     'Pdf',
 ]